refactor(i18n): rename namespace file
