+++ src/i18n/namespaces.ts
